Split things up across three files, templated the socket type for Session.
